+++ doc/mysql-schema/2021-10-30/change.sql
@@ -1,0 +1,1 @@
+ALTER TABLE alert ADD COLUMN feishu_robot_hook VARCHAR(500) NULL COMMENT '飞书机器人hook地址';