+++ frostmourne-monitor/src/main/java/com/autohome/frostmourne/monitor/service/core/alert/AlertService.java
@@ -88,6 +88,7 @@ public class AlertService implements IAlertService {
         alarmMessage.setDingHook(alertContract.getDingRobotHook());
         alarmMessage.setHttpPostEndpoint(alertContract.getHttpPostUrl());
         alarmMessage.setWechatHook(alertContract.getWechatRobotHook());
+        alarmMessage.setFeiShuHook(alertContract.getFeishuRobotHook());
         alarmMessage.setContext(alarmProcessLogger.getContext());
         Protocol<List<MessageResult>> protocol = frostmourneSpiApi.send(alarmMessage, "frostmourne-monitor");
         if (protocol.getReturncode() != 0) {